Removed MongoRepoLive
We just need a MongoRepo class to handle repositories.
Also, MongoRepo is now a `case class` and can be imported in the
custom repositories an an instance value, allowing the custom
repositories to create their own functions and avoiding conflicts
with the MongoRepo functions.

So now instead of this:
```
case class ItemsRepoLive extends MongoRepoLive[Item] {
  override databaseName: String = ""
  override collectionName: String = ""
}
```

we do this:
```
case class ItemsRepoLive(repo: MongoRepo[Item])
```

An example can be found in `core/src/test/scala/zongo/support/ItemsRepo.scala`